Part works before page 110.
Setting up the project
Let’s open Unreal Engine 4! Perform the following steps:
1. First, we have to change our Enemy to act as a wandering frightened monster. So,
double-click on the Enemy AIController. Now, navigate to Event Graph and zoom
in.
2. We want to create a new variable for State and make this Integer.
3. Next, we have to change the Find Hero script. We will adapt this to detect the Hero.
What we also want to keep is the chasing of the Hero blueprint. This will be used to
chase or run from us scared, which is the inverse. So, focus on the script.
4. Remove Event Begin Play from the beginning of the script. Next, pull back Event
Tick from the Chase Hero script.
5. Replace the Delay node that is after the Event Tick node in the Chase Hero script
and put a Switch on Int node. Then, pull the State variable and connect it to the
Selection value pin:
Note
Switch on Int is a very useful function. I’ve used it countless times to replace the
redundant work that you could usually do in passes, such as recurring a task a
number a times before going to the next index.
6. Now, the AI will constantly tick based on State. Different scripts are executed, which
can be transitioned in between, and this is the basis for Finite State Machine.
7. Next, let’s conjoin our two scripts to the Switch on Int node. We want the Find Hero
script to be connected to the index 0 of the node. Then, connect index 1 of the node
to the Chase Hero script. So, you can press Ctrl and click on index 0, moving Move
to Location to index 1 after connecting index 0 to Get All Actors Of Class.
8. So, we have to check the distance of the Actors element found from Get All Actors
Of Class, and then, if an Actor is within range, we will roll the chance of transiting
into Attacking or Running away from this Actor: